motion_detector/requirements.txt
tensorflow>=2.15
keras>=2.15
pandas>=2.0
numpy>=1.24
h5py>=3.8
opencv-python>=4.9
imageio
pillow
matplotlib